implmented fixed point, needs a lot of opt

+++ src/Rendering/renderUtil/interpolation/flat/FlatLerper_Base.java
@@ -1,9 +1,9 @@
 package Rendering.renderUtil.interpolation.flat;
 
 public abstract class FlatLerper_Base {
-    public float z;
-    public float invW;
+    public int z;
+    public int invW;
 
-    public float tex_u, tex_v;
-    public float spec_u, spec_v;
+    public int tex_u, tex_v;
+    public int spec_u, spec_v;
 }
