+++ src/Rendering/renderUtil/interpolation/flat/FlatLerper_R.java
@@ -5,11 +5,11 @@ public class FlatLerper_R extends FlatLerper_Base {
     public FlatLerper_R() {
     }
 
-    public FlatLerper_R(float z, float invW, float tex_u, float tex_v, float spec_u, float spec_v) {
+    public FlatLerper_R(int z, int invW, int tex_u, int tex_v, int spec_u, int spec_v) {
         reset(z, invW, tex_u, tex_v, spec_u, spec_v);
     }
 
-    public void reset(float z, float invW, float tex_u, float tex_v, float spec_u, float spec_v) {
+    public void reset(int z, int invW, int tex_u, int tex_v, int spec_u, int spec_v) {
         this.z = z;
         this.invW = invW;
         this.tex_u = tex_u;
