+++ src/Rendering/renderUtil/interpolation/phong/PhongLerper_Base.java
@@ -1,12 +1,12 @@
 package Rendering.renderUtil.interpolation.phong;
 
 public abstract class PhongLerper_Base {
-    public float x, z;
-    public float invW;
+    public int x, z;
+    public int invW;
 
-    public float tex_u, tex_v;
-    public float spec_u, spec_v;
+    public int tex_u, tex_v;
+    public int spec_u, spec_v;
 
-    public float p_ws_x, p_ws_y, p_ws_z;
-    public float n_ws_x, n_ws_y, n_ws_z;
+    public int p_ws_x, p_ws_y, p_ws_z;
+    public int n_ws_x, n_ws_y, n_ws_z;
 }
